Import the latest Interridge Vents Field data

Import the latest version of the csv file from the InterRidge Vents Database https://vents-data.interridge.org/ventfields_list_all, presently in Version 3.4. The function imports the latest version into data-raw. Column names are regularised by replacing spaces with underscores,removing brackets and a download_date column is added. The file is saved to data as vents.rda and loaded into the global environment as vents.

import_vents()

Value

data.frame

Details

For license details see interridge_citation.txt in data-raw.

References

Please cite the data as: "Beaulieu, S.E., 2015, InterRidge Global Database of Active Submarine Hydrothermal Vent Fields: prepared for InterRidge, Version 3.3. World Wide Web electronic publication. Version 3.X accessed 2017-09-04, http://vents-data.interridge.org"

Examples

if (FALSE) vents <- import_vents()