vents

A dataset on global hydrothermal vents from the InterRidge Vents Database Version 3.4 downloaded from https://vents-data.interridge.org/ventfields_list_all

vents

Format

A data.frame with 26 variables:

name_id

character vector

name_aliases

character vector

mgds_featureid_lowest_in_hierarchy

character vector

vent_sites

character vector

activity

character vector

maximum_temperature

integer vector

max_temperature_category

character vector

latitude

double

longitude

double

ocean

character vector

region

character vector

national_jurisdiction

character vector

maximum_or_single_reported_depth

integer

minimum_depth

integer

tectonic_setting

character vector

full_spreading_rate_mma

double

host_rock

character vector

deposit_type

character vector

notes_on_vent_field_description

character vector

notes_relevant_to_biology

character vector

year_and_how_discovered_if_active,_visual_confirmation_is_listed_first

character vector

discovery_references_text

character vector

other_references_text

character vector

volcano_number

integer

node_id

integer

download_date

date

Details

For license details see interridge_citation.txt in data-raw or https://creativecommons.org/licenses/by-nc-sa/3.0/.

References

Please cite the data as: "Beaulieu, S.E., 2015, InterRidge Global Database of Active Submarine Hydrothermal Vent Fields: prepared for InterRidge, Version 3.3. World Wide Web electronic publication. Version 3.X accessed, insert latest date as YYYY-MM-DD, http://vents-data.interridge.org"